requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
dateparser==1.0.0
firebase==3.0.1
Flask==1.1.2
gcloud==0.18.3
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
jwcrypto==0.8
MarkupSafe==1.1.1
multidict==5.1.0
oauth2client==4.1.3
protobuf==3.16.0rc1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
pyparsing==2.4.7
python-binance==1.0.1
python-dateutil==2.8.1
python-jwt==3.3.0
pytz==2021.1
regex==2021.4.4
requests==2.25.1
requests-toolbelt==0.9.1
rsa==4.7.2
six==1.15.0
sseclient==0.0.27
typing-extensions==3.7.4.3
tzlocal==2.1
ujson==4.0.2
urllib3==1.26.4
websockets==8.1
Werkzeug==1.0.1
yarl==1.6.3
